+++ CHANGELOG.md
@@ -24,3 +24,4 @@ that users understand how the changes affect the new version.
 ### Changes
 - Add support for PEP sample configuration
 - Add support for Snakemake 7.8.5
+- Remove variants plots